+++ eras/alonzo/impl/testlib/Test/Cardano/Ledger/Alonzo/Examples.hs
@@ -1,16 +1,19 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
 
+-- | The example transactions in this module are not valid transactions. We
+-- don't care, we are only interested in serialisation, not validation.
 module Test.Cardano.Ledger.Alonzo.Examples (
   ledgerExamples,
   mkAlonzoBasedLedgerExamples,
-  mkAlonzoBasedExampleTx,
-  exampleAlonzoBasedShelleyTxBody,
-  exampleAlonzoBasedTxBody,
+  exampleAlonzoBasedTx,
+  exampleAlonzoBasedTopTx,
+  addAlonzoToConwayExampleReqSigners,
   exampleDatum,
 ) where
 
@@ -19,7 +22,6 @@ import Cardano.Ledger.Alonzo.Core
 import Cardano.Ledger.Alonzo.Genesis (AlonzoGenesis (..))
 import Cardano.Ledger.Alonzo.Plutus.Context (EraPlutusTxInfo)
 import Cardano.Ledger.Alonzo.Scripts (
-  AlonzoPlutusPurpose (..),
   ExUnits (..),
   Prices (..),
  )
@@ -44,22 +46,26 @@ import Cardano.Ledger.Shelley.Rules (
 import Data.Default (def)
 import qualified Data.List.NonEmpty as NE
 import qualified Data.Map.Strict as Map
+import qualified Data.MapExtras as Map
 import qualified Data.Sequence.Strict as StrictSeq
 import qualified Data.Set as Set
+import Data.Typeable (Typeable)
 import Lens.Micro
 import qualified PlutusLedgerApi.Common as P
-import Test.Cardano.Ledger.Allegra.Examples (mkAllegraBasedExampleTx)
 import Test.Cardano.Ledger.Alonzo.Arbitrary (alwaysSucceeds)
 import Test.Cardano.Ledger.Core.KeyPair (mkAddr)
 import Test.Cardano.Ledger.Core.Utils (mkDummySafeHash, unsafeBoundRational)
 import Test.Cardano.Ledger.Mary.Examples (
-  exampleMaryBasedShelleyTxBody,
-  exampleMaryBasedTxBody,
+  exampleMaryBasedTx,
   exampleMultiAssetValue,
  )
 import Test.Cardano.Ledger.Plutus (alwaysFailsPlutus, zeroTestingCostModelV1)
 import Test.Cardano.Ledger.Shelley.Examples (
   LedgerExamples (..),
+  addShelleyBasedTopTxExampleFee,
+  addShelleyToBabbageExampleProposedPUpdates,
+  addShelleyToBabbageTxCerts,
+  addShelleyToConwayTxCerts,
   exampleNewEpochState,
   exampleNonMyopicRewards,
   examplePayKey,
@@ -82,8 +88,29 @@ ledgerExamples =
                 DelegateeNotRegisteredDELEG @AlonzoEra (mkKeyHash 1)
     )
     exampleAlonzoNewEpochState
-    exampleTxAlonzo
+    exampleAlonzoTx
     exampleAlonzoGenesis
+  where
+    exampleAlonzoTx =
+      exampleAlonzoBasedTopTx
+        & addShelleyBasedTopTxExampleFee
+        & addShelleyToBabbageExampleProposedPUpdates
+        & addShelleyToBabbageTxCerts
+        & addShelleyToConwayTxCerts
+        & addAlonzoToConwayExampleReqSigners
+
+    exampleAlonzoGenesis =
+      AlonzoGenesis
+        { agCoinsPerUTxOWord = CoinPerWord $ Coin 1
+        , agPlutusV1CostModel = zeroTestingCostModelV1
+        , agPrices = Prices (unsafeBoundRational 90) (unsafeBoundRational 91)
+        , agMaxTxExUnits = ExUnits 123 123
+        , agMaxBlockExUnits = ExUnits 223 223
+        , agMaxValSize = 1234
+        , agCollateralPercentage = 20
+        , agMaxCollateralInputs = 30
+        , agExtraConfig = Nothing
+        }
 
 mkAlonzoBasedLedgerExamples ::
   forall era.
@@ -127,110 +154,95 @@ exampleAlonzoNewEpochState =
     emptyPParams
     (emptyPParams & ppCoinsPerUTxOWordL .~ CoinPerWord (Coin 1))
 
-exampleTxAlonzo :: Tx TopTx AlonzoEra
-exampleTxAlonzo =
-  mkAlonzoBasedExampleTx
-    exampleAlonzoBasedShelleyTxBody
-    (AlonzoSpending $ AsIx 0)
-
-mkAlonzoBasedExampleTx ::
+exampleAlonzoBasedTopTx ::
   forall era.
   ( AlonzoEraTx era
   , AlonzoEraTxAuxData era
   , EraPlutusTxInfo 'PlutusV1 era
+  , Value era ~ MaryValue
   ) =>
-  TxBody TopTx era ->
-  PlutusPurpose AsIx era ->
   Tx TopTx era
-mkAlonzoBasedExampleTx txBody scriptPurpose =
-  mkAllegraBasedExampleTx
-    txBody
+exampleAlonzoBasedTopTx =
+  exampleAlonzoBasedTx
+    & addAlonzoBasedTopTxFeatureExamples
+
+exampleAlonzoBasedTx ::
+  forall era l.
+  ( AlonzoEraTx era
+  , AlonzoEraTxAuxData era
+  , EraPlutusTxInfo 'PlutusV1 era
+  , Value era ~ MaryValue
+  , Typeable l
+  ) =>
+  Tx l era
+exampleAlonzoBasedTx =
+  exampleMaryBasedTx
+    & addAlonzoBasedTxFeatureExamples
+
+addAlonzoBasedTopTxFeatureExamples ::
+  forall era.
+  AlonzoEraTx era =>
+  Tx TopTx era ->
+  Tx TopTx era
+addAlonzoBasedTopTxFeatureExamples tx =
+  tx
+    & isValidTxL .~ IsValid True
+    & bodyTxL . collateralInputsTxBodyL <>~ exampleTxIns
+
+addAlonzoBasedTxFeatureExamples ::
+  forall era l.
+  ( AlonzoEraTx era
+  , AlonzoEraTxAuxData era
+  , EraPlutusTxInfo 'PlutusV1 era
+  , Value era ~ MaryValue
+  ) =>
+  Tx l era ->
+  Tx l era
+addAlonzoBasedTxFeatureExamples tx =
+  tx
     & witsTxL
       <>~ ( mkBasicTxWits
-              & scriptTxWitsL
-                .~ Map.singleton
-                  (hashScript @era $ alwaysSucceeds @'PlutusV1 3)
-                  (alwaysSucceeds @'PlutusV1 3)
-              & datsTxWitsL .~ TxDats (Map.singleton (hashData (exampleDatum @era)) exampleDatum)
-              & rdmrsTxWitsL
-                .~ Redeemers (Map.singleton scriptPurpose (exampleRedeemer, ExUnits 5000 5000))
+              & scriptTxWitsL <>~ Map.fromElems hashScript [alwaysSucceeds @'PlutusV1 3]
+              & datsTxWitsL <>~ TxDats (Map.fromElems hashData [exampleDatum])
+              & rdmrsTxWitsL <>~ redeemers
           )
-    & auxDataTxL
-      %~ fmap
-        ( \auxData ->
-            auxData
-              & plutusScriptsTxAuxDataL
-                <>~ Map.singleton PlutusV1 (NE.singleton (plutusBinary (alwaysFailsPlutus @'PlutusV1 2)))
-        )
-    & isValidTxL .~ IsValid True
-
-exampleAlonzoBasedTxBody ::
-  forall era.
-  ( AlonzoEraTxBody era
-  , Value era ~ MaryValue
-  ) =>
-  TxBody TopTx era
-exampleAlonzoBasedTxBody =
-  mkAlonzoBasedExampleTxBody exampleMaryBasedTxBody
-
-exampleAlonzoBasedShelleyTxBody ::
-  forall era.
-  ( AlonzoEraTxBody era
-  , ShelleyEraTxBody era
-  , Value era ~ MaryValue
-  ) =>
-  TxBody TopTx era
-exampleAlonzoBasedShelleyTxBody =
-  mkAlonzoBasedPreConwayExampleTxBody exampleMaryBasedShelleyTxBody
-
-mkAlonzoBasedPreConwayExampleTxBody ::
-  forall era.
-  ( AlonzoEraTxBody era
-  , Value era ~ MaryValue
-  , AtMostEra "Conway" era
-  ) =>
-  TxBody TopTx era ->
-  TxBody TopTx era
-mkAlonzoBasedPreConwayExampleTxBody txBody =
-  mkAlonzoBasedExampleTxBody txBody
-    & reqSignerHashesTxBodyL .~ Set.singleton (mkKeyHash 212)
-
-mkAlonzoBasedExampleTxBody ::
-  forall era.
-  ( AlonzoEraTxBody era
-  , Value era ~ MaryValue
-  ) =>
-  TxBody TopTx era ->
-  TxBody TopTx era
-mkAlonzoBasedExampleTxBody txBody =
-  txBody
-    & collateralInputsTxBodyL .~ exampleTxIns
-    & outputsTxBodyL
+    & modifyTxAuxData
+      ( plutusScriptsTxAuxDataL
+          %~ Map.insertWith
+            (<>)
+            PlutusV1
+            (NE.singleton (plutusBinary (alwaysFailsPlutus @'PlutusV1 2)))
+      )
+    & bodyTxL . outputsTxBodyL
       <>~ StrictSeq.fromList
         [ mkBasicTxOut
             (mkAddr examplePayKey exampleStakeKey)
             (exampleMultiAssetValue 3)
             & dataHashTxOutL .~ SJust (mkDummySafeHash 1)
         ]
-    & scriptIntegrityHashTxBodyL .~ SJust (mkDummySafeHash 42)
-    & networkIdTxBodyL .~ SJust Mainnet
+    & bodyTxL . scriptIntegrityHashTxBodyL .~ SJust (mkDummySafeHash 42)
+    & bodyTxL . networkIdTxBodyL .~ SJust Mainnet
+  where
+    redeemers =
+      Redeemers $
+        Map.fromList
+          [ (SpendingPurpose $ AsIx 0, (redeemerData, ExUnits 5000 5000))
+          , (MintingPurpose $ AsIx 1, (redeemerData, ExUnits 5000 5000))
+          , (CertifyingPurpose $ AsIx 2, (redeemerData, ExUnits 5000 5000))
+          , (RewardingPurpose $ AsIx 3, (redeemerData, ExUnits 5000 5000))
+          ]
+    redeemerData = Data (P.Constr 1 [P.List [P.I 1], P.Map [(P.I 2, P.B "2")]])
+
+addAlonzoToConwayExampleReqSigners ::
+  forall era l.
+  ( AlonzoEraTxBody era
+  , AtMostEra "Conway" era
+  , EraTx era
+  ) =>
+  Tx l era ->
+  Tx l era
+addAlonzoToConwayExampleReqSigners tx =
+  tx & bodyTxL . reqSignerHashesTxBodyL <>~ Set.singleton (mkKeyHash 212)
 
 exampleDatum :: Era era => Data era
-exampleDatum = Data (P.I 191)
-
-exampleRedeemer :: Era era => Data era
-exampleRedeemer = Data (P.I 919)
-
-exampleAlonzoGenesis :: AlonzoGenesis
-exampleAlonzoGenesis =
-  AlonzoGenesis
-    { agCoinsPerUTxOWord = CoinPerWord $ Coin 1
-    , agPlutusV1CostModel = zeroTestingCostModelV1
-    , agPrices = Prices (unsafeBoundRational 90) (unsafeBoundRational 91)
-    , agMaxTxExUnits = ExUnits 123 123
-    , agMaxBlockExUnits = ExUnits 223 223
-    , agMaxValSize = 1234
-    , agCollateralPercentage = 20
-    , agMaxCollateralInputs = 30
-    , agExtraConfig = Nothing
-    }
+exampleDatum = Data (P.Constr 0 [P.List [P.I 0], P.Map [(P.I 1, P.B "1")]])