+++ eras/alonzo/impl/CHANGELOG.md
@@ -43,6 +43,9 @@
 * Add `genDatumPresent`
 * Added `Proxy era` argument to `exUnitsRule`
 * `TranslationInstance` has a new field `tiPlutusPurpose`
+* In `Test.Cardano.Ledger.Alonzo.Examples`:
+  - Remove `mkAlonzoBasedExampleTx`, `exampleAlonzoBasedShelleyTxBody`, `exampleAlonzoBasedTxBody`, `exampleRedeemer`
+  - Add `exampleAlonzoBasedTopTx`, `exampleAlonzoBasedTopTx`, `addAlonzoToConwayExampleReqSigners`
 
 ## 1.15.0.0